Fix clippy

+++ src/math.rs
@@ -193,7 +193,7 @@ pub fn floor_sum(n: i64, m: i64, a: i64, b: i64) -> i64 {
     use std::num::Wrapping as W;
     assert!((0..1i64 << 32).contains(&n));
     assert!((1..1i64 << 32).contains(&m));
-    let mut ans = W(0 as u64);
+    let mut ans = W(0_u64);
     let (wn, wm, mut wa, mut wb) = (W(n as u64), W(m as u64), W(a as u64), W(b as u64));
     if a < 0 {
         let a2 = W(internal_math::safe_mod(a, m) as u64);
